+++ src/cmd/separate.rs
@@ -24,7 +24,8 @@ to some splitting algorithm that can be one of:
     * --offsets: extract byte slices
 
 Created columns can be given a name using the --into flag, else they will be
-given generic names like "split1",  "split2" and so forth.
+given generic names based on the original column name. For instance, splitting a
+column named "text" will produce columns named "text1", "text2"...
 
 It is also possible to limit the number of splits using the --max flag.
 
@@ -91,9 +92,15 @@ separate options:
                            By default, all possible splits are made.
     --into <column-names>  Specify names for the new columns created by the
                            splits. If not provided, new columns will be named
-                           split1, split2, etc. If used with --max,
+                           before the original column name ('text' column will
+                           be separated into 'text1', 'text2', etc.). If used with --max,
                            the number of names provided must be equal or lower
-                           than <n>.
+                           than <n>. Cannot be used with --prefix.
+    --prefix <prefix>      Specify a prefix for the new columns created by the
+                           splits. By default, no prefix is used and new columns
+                           are named before the original column name ('text'
+                           column will be separated into 'text1', 'text2', etc.).
+                           Cannot be used with --into.
     --too-many <option>    Specify how to handle extra cells when the number
                            of splitted cells exceeds --max, or
                            the number of provided names with --into.
@@ -120,7 +127,7 @@ Common options:
 
 #[derive(Deserialize, Debug)]
 struct Args {
-    arg_column: SelectedColumns,
+    arg_column: String,
     arg_separator: String,
     arg_input: Option<String>,
     flag_regex: bool,
@@ -129,6 +136,7 @@ struct Args {
     flag_keep: bool,
     flag_max: Option<usize>,
     flag_into: Option<String>,
+    flag_prefix: Option<String>,
     flag_too_many: TooManyMode,
     flag_output: Option<String>,
     flag_no_headers: bool,
@@ -363,16 +371,22 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
         .as_ref()
         .map(|names| util::str_to_csv_byte_record(names));
 
-    match (args.flag_max, &new_column_names) {
-        (Some(max), Some(names)) if names.len() > max => {
+    match (args.flag_max, &new_column_names, &args.flag_prefix) {
+        (_, Some(_), Some(_)) => {
+            Err("--into and --prefix cannot be used together!")?;
+        }
+        (Some(max), Some(names), _) if names.len() > max => {
             Err(format!("--into cannot specify more column names than --max : got {} for --into and {} for --max", names.len(), max))?;
         }
         _ => (),
     }
 
+    let column_to_separate_name = args.arg_column;
+    let column_to_separate_sel = SelectedColumns::try_from(column_to_separate_name.clone())?;
+
     let rconf = Config::new(&args.arg_input)
         .no_headers(args.flag_no_headers)
-        .select(args.arg_column)
+        .select(column_to_separate_sel)
         .delimiter(args.flag_delimiter);
 
     let mut wtr = Config::new(&args.flag_output).simd_writer()?;
@@ -489,6 +503,7 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
         );
 
         let mut number_of_new_columns = max_splits;
+        let prefix = args.flag_prefix.unwrap_or(column_to_separate_name);
 
         if let Some(names) = &new_column_names {
             new_headers.extend(names);
@@ -496,7 +511,7 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
         }
 
         for i in 1..=number_of_new_columns {
-            let header_name = format!("split{}", i);
+            let header_name = format!("{}{}", prefix, i);
             new_headers.push_field(header_name.as_bytes());
         }