+++ src/main.rs
@@ -85,6 +85,7 @@ macro_rules! command_list {
     enum        Enumerate CSV file by preprending an index column
     flatmap     Emit one row per value yielded by an expression evaluated for each CSV row
     fill        Fill empty cells
+    complete    Add missing rows in a column of contiguous values
     blank       Blank down contiguous identical cell values
     separate    Split a single column into multiple ones
 
@@ -103,8 +104,8 @@ macro_rules! command_list {
 
 ## Transpose & pivot
     transpose (t)            Transpose CSV file
-    pivot     (pivot-wider)  Stack multiple columns into fewer columns
-    unpivot   (pivot-longer) Split distinct values of a column into their own columns
+    pivot     (pivot-wider)  Split distinct values of a column into their own columns
+    unpivot   (pivot-longer) Stack multiple columns into fewer columns
 
 ## Split a CSV file into multiple
     split       Split CSV data into chunks
@@ -277,6 +278,7 @@ enum Command {
     Cat,
     Cluster,
     Compgen,
+    Complete,
     Completions,
     Count,
     Dedup,
@@ -367,6 +369,7 @@ impl Command {
                 cmd::compgen::run();
                 Ok(())
             }
+            Command::Complete => cmd::complete::run(argv),
             Command::Completions => cmd::completions::run(argv),
             Command::Count => cmd::count::run(argv),
             Command::Dedup => cmd::dedup::run(argv),