+++ src/moonblade/functions.rs
@@ -79,11 +79,12 @@ pub fn get_function(name: &str) -> Option<(Function, FunctionArguments)> {
             |args| argcompare(args, Ordering::is_lt),
             FunctionArguments::with_range(1..=2),
         ),
+        "basename" => (basename, FunctionArguments::with_range(1..=2)),
         "bytesize" => (bytesize, FunctionArguments::unary()),
         "carry_stemmer" => (carry_stemmer_fn, FunctionArguments::unary()),
         "ceil" => (
-            |args| unary_arithmetic_op(args, DynamicNumber::ceil),
-            FunctionArguments::unary(),
+            |args| round_like_op(args, DynamicNumber::ceil),
+            FunctionArguments::with_range(1..=2),
         ),
         "cmd" => (cmd, FunctionArguments::binary()),
         "compact" => (compact, FunctionArguments::unary()),
@@ -99,6 +100,7 @@ pub fn get_function(name: &str) -> Option<(Function, FunctionArguments)> {
                 Argument::with_name("timezone"),
             ]),
         ),
+        "dirname" => (dirname, FunctionArguments::unary()),
         "div" => (
             |args| variadic_arithmetic_op(args, Div::div),
             FunctionArguments::variadic(2),
@@ -122,8 +124,8 @@ pub fn get_function(name: &str) -> Option<(Function, FunctionArguments)> {
         "first" => (first, FunctionArguments::unary()),
         "float" => (parse_float, FunctionArguments::unary()),
         "floor" => (
-            |args| unary_arithmetic_op(args, DynamicNumber::floor),
-            FunctionArguments::unary(),
+            |args| round_like_op(args, DynamicNumber::floor),
+            FunctionArguments::with_range(1..=2),
         ),
         "fmt" => (fmt, FunctionArguments::variadic(2)),
         "numfmt" => (
@@ -224,6 +226,7 @@ pub fn get_function(name: &str) -> Option<(Function, FunctionArguments)> {
         ),
         "parse_dataurl" => (parse_dataurl, FunctionArguments::unary()),
         "parse_json" => (parse_json, FunctionArguments::unary()),
+        "parse_py_literal" => (parse_py_literal, FunctionArguments::unary()),
         "pjoin" | "pathjoin" => (pathjoin, FunctionArguments::variadic(2)),
         "pow" => (
             |args| binary_arithmetic_op(args, DynamicNumber::pow),
@@ -244,8 +247,8 @@ pub fn get_function(name: &str) -> Option<(Function, FunctionArguments)> {
         "regex" => (parse_regex, FunctionArguments::unary()),
         "replace" => (replace, FunctionArguments::nary(3)),
         "round" => (
-            |args| unary_arithmetic_op(args, DynamicNumber::round),
-            FunctionArguments::unary(),
+            |args| round_like_op(args, DynamicNumber::round),
+            FunctionArguments::with_range(1..=2),
         ),
         "shell" => (shell, FunctionArguments::unary()),
         "shlex_split" => (shlex_split, FunctionArguments::unary()),
@@ -303,8 +306,8 @@ pub fn get_function(name: &str) -> Option<(Function, FunctionArguments)> {
         "ltrim" => (ltrim, FunctionArguments::with_range(1..=2)),
         "rtrim" => (rtrim, FunctionArguments::with_range(1..=2)),
         "trunc" => (
-            |args| unary_arithmetic_op(args, DynamicNumber::trunc),
-            FunctionArguments::unary(),
+            |args| round_like_op(args, DynamicNumber::trunc),
+            FunctionArguments::with_range(1..=2),
         ),
         "typeof" => (type_of, FunctionArguments::unary()),
         "unidecode" => (apply_unidecode, FunctionArguments::unary()),
@@ -1089,6 +1092,22 @@ where
     Ok(DynamicValue::from(op(args.pop1_number()?)))
 }
 
+fn round_like_op<F>(mut args: BoundArguments, op: F) -> FunctionResult
+where
+    F: Fn(DynamicNumber) -> DynamicNumber,
+{
+    if args.len() == 2 {
+        let unit = args.pop1_number()?;
+        let operand = args.pop1_number()?;
+
+        let result = op(operand / unit) * unit;
+
+        Ok(DynamicValue::from(result))
+    } else {
+        Ok(DynamicValue::from(op(args.pop1_number()?)))
+    }
+}
+
 fn binary_arithmetic_op<F>(args: BoundArguments, op: F) -> FunctionResult
 where
     F: Fn(DynamicNumber, DynamicNumber) -> DynamicNumber,
@@ -1476,11 +1495,36 @@ fn copy_file(args: BoundArguments) -> FunctionResult {
 }
 
 fn ext(args: BoundArguments) -> FunctionResult {
-    let path = PathBuf::from(args.get1_str()?.as_ref());
+    let string = args.get1_str()?;
+    let path = Path::new(string.as_ref());
 
     Ok(DynamicValue::from(
         path.extension().and_then(|e| e.to_str()),
     ))
+}
+
+fn dirname(args: BoundArguments) -> FunctionResult {
+    let string = args.get1_str()?;
+    let path = Path::new(string.as_ref());
+
+    Ok(DynamicValue::from(path.parent().and_then(|p| p.to_str())))
+}
+
+fn basename(args: BoundArguments) -> FunctionResult {
+    let string = args.get1_str()?;
+    let path = Path::new(string.as_ref());
+
+    let name = path.file_name().and_then(|p| p.to_str());
+
+    if args.len() == 2 {
+        let suffix = args.get(1).unwrap().try_as_str()?;
+
+        Ok(DynamicValue::from(
+            name.and_then(|n| n.strip_suffix(suffix.as_ref()).or(Some(n))),
+        ))
+    } else {
+        Ok(DynamicValue::from(name))
+    }
 }
 
 fn filesize(args: BoundArguments) -> FunctionResult {
@@ -1682,6 +1726,55 @@ fn parse_json(args: BoundArguments) -> FunctionResult {
 
     serde_json::from_slice(arg.try_as_bytes()?)
         .map_err(|_| EvaluationError::JSONParseError(format!("{:?}", args.get1())))
+}
+
+fn parse_py_literal(args: BoundArguments) -> FunctionResult {
+    let parsed: py_literal::Value = args
+        .get1_str()?
+        .parse()
+        .map_err(|err: py_literal::ParseError| EvaluationError::Custom(err.to_string()))?;
+
+    fn map_to_dynamic_value(value: py_literal::Value) -> FunctionResult {
+        Ok(match value {
+            py_literal::Value::None => DynamicValue::None,
+            py_literal::Value::Boolean(v) => DynamicValue::Boolean(v),
+            py_literal::Value::Float(f) => DynamicValue::Float(f),
+            py_literal::Value::Integer(bi) => match bi.try_into() {
+                Ok(i) => DynamicValue::Integer(i),
+                Err(err) => return Err(EvaluationError::Custom(err.to_string())),
+            },
+            py_literal::Value::Bytes(b) => DynamicValue::from_owned_bytes(b),
+            py_literal::Value::String(s) => DynamicValue::from(s),
+            py_literal::Value::List(l)
+            | py_literal::Value::Tuple(l)
+            | py_literal::Value::Set(l) => {
+                let mut list = Vec::new();
+
+                for item in l {
+                    list.push(map_to_dynamic_value(item)?);
+                }
+
+                DynamicValue::from(list)
+            }
+            py_literal::Value::Dict(d) => {
+                let mut dict = HashMap::new();
+
+                for (key, value) in d {
+                    dict.insert(
+                        map_to_dynamic_value(key)?.try_as_str()?.into_owned(),
+                        map_to_dynamic_value(value)?,
+                    );
+                }
+
+                DynamicValue::from(dict)
+            }
+            py_literal::Value::Complex(c) => {
+                DynamicValue::from(vec![DynamicValue::Float(c.re), DynamicValue::Float(c.im)])
+            }
+        })
+    }
+
+    map_to_dynamic_value(parsed)
 }
 
 fn parse_dataurl(args: BoundArguments) -> FunctionResult {